Track distance to location/s

Calculate geodesic distance to a location or locations based on longitude, latitude (from) input vectors and longitude, latitude (to) input vectors. The unit of distance is metres. The to values may be a single value or individual to each from location.

track_distance_to(x, y, to_x, to_y)

Arguments

x

longitude

y

latitude

to_x

longitude vector of to location/s

to_y

latitude vector of to locations/s

Value

a numeric vector of distance-to values in metres

Details

No missing values are required as padding, but input data with NAs will incur an NA in the output.

To use this on multiple track ids, use a grouped data frame with tidyverse code like data %>% group_by(id) %>% mutate(distance = track_distance_to(lon, lat, to_lon, to_lat))

Examples

track_distance_to(trips0$x, trips0$y, to_x = 147, to_y = -42)[1:10]
#>  [1] 2605548 2554214 2575812 2509483 2519693 2422175 2474630 2433576 2436278
#> [10] 2399168