Track intermediate points

Calculate great circle intermediate points on longitude, latitude input vectors. A spherical model is used, from the geosphere package.

track_intermediate(x, y, date = NULL, distance = NULL,
  duration = NULL)

Arguments

x
longitude

y
latitude

date
optional input date-time in POSIXct

distance
optional minimum distance (metres) between interpolated points

duration
optional minimum duration (seconds) between interpolated point, if set then distance must be NULL and date must be input

Value

a list of data frames of intermediate points (for use with unnest() from tidyr)

Details

This function returns a list of data frames, with a data frame of interpolated locations for every interval between input locations. There is a final empty data frame to ensure the list is the same length as the inputs. See embedded usage of the tidyr function 'unnest()' for ease of use.

To use on multiple track ids, use a grouped data frame with tidyverse code like inter <- data %>% group_by(id) %>% mutate(inter = track_intermediate(lon, lat, date = , distance = ).

Then, un-nest this result for further use (the 'inter' above retains the information about the parent locations for custom usage if needed), so the final location of each group has invalid intermediates: dd <- inter %>% slice(-1) %>% unnest()

Examples

track_intermediate(trips0$x[1:10], trips0$y[1:10], distance = 15000)

#> $`1`
#>      int_x     int_y
#> 1 115.2767 -42.37580
#> 2 115.3878 -42.25572
#> 3 115.4985 -42.13554
#> 4 115.6087 -42.01526
#> 5 115.7186 -41.89487
#> 6 115.8281 -41.77437
#> 7 115.9371 -41.65377
#> 8 116.0457 -41.53307
#> 9 116.1539 -41.41227
#> 
#> $`2`
#>      int_x     int_y
#> 1 116.1539 -41.41227
#> 2 116.0618 -41.27267
#> 3 115.9701 -41.13299
#> 
#> $`3`
#>      int_x     int_y
#> 1 115.9701 -41.13299
#> 2 116.0372 -41.25886
#> 3 116.1046 -41.38469
#> 4 116.1723 -41.51048
#> 5 116.2402 -41.63622
#> 6 116.3084 -41.76193
#> 7 116.3769 -41.88760
#> 8 116.4456 -42.01323
#> 9 116.5146 -42.13881
#> 
#> $`4`
#>      int_x     int_y
#> 1 116.5146 -42.13881
#> 2 116.4852 -42.00301
#> 3 116.4558 -41.86720
#> 
#> $`5`
#>      int_x     int_y
#> 1 116.4558 -41.86720
#> 2 116.6497 -41.88856
#> 3 116.8436 -41.90960
#> 4 117.0377 -41.93031
#> 5 117.2320 -41.95069
#> 6 117.4263 -41.97074
#> 7 117.6208 -41.99047
#> 
#> $`6`
#>      int_x     int_y
#> 1 117.6208 -41.99047
#> 2 117.4294 -41.89645
#> 3 117.2385 -41.80212
#> 4 117.0481 -41.70748
#> 
#> $`7`
#>      int_x     int_y
#> 1 117.0481 -41.70748
#> 2 117.1528 -41.59157
#> 3 117.2570 -41.47557
#> 4 117.3609 -41.35948
#> 5 117.4643 -41.24329
#> 6 117.5675 -41.12701
#> 7 117.6702 -41.01064
#> 8 117.7726 -40.89418
#> 
#> $`8`
#>      int_x     int_y
#> 1 117.7726 -40.89418
#> 2 117.8157 -40.75087
#> 3 117.8586 -40.60754
#> 4 117.9014 -40.46420
#> 5 117.9439 -40.32084
#> 6 117.9863 -40.17746
#> 7 118.0285 -40.03407
#> 8 118.0705 -39.89067
#> 9 118.1123 -39.74725
#> 
#> $`9`
#>      int_x     int_y
#> 1 118.1123 -39.74725
#> 2 118.1511 -39.89936
#> 3 118.1900 -40.05146
#> 4 118.2292 -40.20355
#> 5 118.2684 -40.35563
#> 6 118.3079 -40.50769
#> 
#> [[10]]
#> data frame with 0 columns and 0 rows
#> 

track_intermediate(trips0$x[1:10], trips0$y[1:10], date = trips0$date,
                                 distance = 1500)

#> $`1`
#>       int_x     int_y            int_date
#> 1  115.2767 -42.37580 2001-01-01 15:39:50
#> 2  115.2870 -42.36463 2001-01-01 15:41:39
#> 3  115.2974 -42.35347 2001-01-01 15:43:29
#> 4  115.3077 -42.34230 2001-01-01 15:45:18
#> 5  115.3180 -42.33113 2001-01-01 15:47:08
#> 6  115.3284 -42.31996 2001-01-01 15:48:57
#> 7  115.3387 -42.30879 2001-01-01 15:50:47
#> 8  115.3491 -42.29762 2001-01-01 15:52:36
#> 9  115.3594 -42.28645 2001-01-01 15:54:26
#> 10 115.3697 -42.27528 2001-01-01 15:56:16
#> 11 115.3800 -42.26410 2001-01-01 15:58:05
#> 12 115.3904 -42.25293 2001-01-01 15:59:55
#> 13 115.4007 -42.24175 2001-01-01 16:01:44
#> 14 115.4110 -42.23058 2001-01-01 16:03:34
#> 15 115.4213 -42.21940 2001-01-01 16:05:23
#> 16 115.4316 -42.20822 2001-01-01 16:07:13
#> 17 115.4419 -42.19704 2001-01-01 16:09:02
#> 18 115.4522 -42.18586 2001-01-01 16:10:52
#> 19 115.4625 -42.17468 2001-01-01 16:12:42
#> 20 115.4728 -42.16350 2001-01-01 16:14:31
#> 21 115.4830 -42.15232 2001-01-01 16:16:21
#> 22 115.4933 -42.14114 2001-01-01 16:18:10
#> 23 115.5036 -42.12995 2001-01-01 16:20:00
#> 24 115.5139 -42.11877 2001-01-01 16:21:49
#> 25 115.5242 -42.10758 2001-01-01 16:23:39
#> 26 115.5344 -42.09639 2001-01-01 16:25:28
#> 27 115.5447 -42.08520 2001-01-01 16:27:18
#> 28 115.5549 -42.07402 2001-01-01 16:29:08
#> 29 115.5652 -42.06283 2001-01-01 16:30:57
#> 30 115.5755 -42.05163 2001-01-01 16:32:47
#> 31 115.5857 -42.04044 2001-01-01 16:34:36
#> 32 115.5959 -42.02925 2001-01-01 16:36:26
#> 33 115.6062 -42.01806 2001-01-01 16:38:15
#> 34 115.6164 -42.00686 2001-01-01 16:40:05
#> 35 115.6267 -41.99567 2001-01-01 16:41:54
#> 36 115.6369 -41.98447 2001-01-01 16:43:44
#> 37 115.6471 -41.97327 2001-01-01 16:45:34
#> 38 115.6573 -41.96208 2001-01-01 16:47:23
#> 39 115.6676 -41.95088 2001-01-01 16:49:13
#> 40 115.6778 -41.93968 2001-01-01 16:51:02
#> 41 115.6880 -41.92848 2001-01-01 16:52:52
#> 42 115.6982 -41.91727 2001-01-01 16:54:41
#> 43 115.7084 -41.90607 2001-01-01 16:56:31
#> 44 115.7186 -41.89487 2001-01-01 16:58:21
#> 45 115.7288 -41.88366 2001-01-01 17:00:10
#> 46 115.7390 -41.87246 2001-01-01 17:02:00
#> 47 115.7492 -41.86125 2001-01-01 17:03:49
#> 48 115.7594 -41.85004 2001-01-01 17:05:39
#> 49 115.7696 -41.83884 2001-01-01 17:07:28
#> 50 115.7797 -41.82763 2001-01-01 17:09:18
#> 51 115.7899 -41.81642 2001-01-01 17:11:07
#> 52 115.8001 -41.80521 2001-01-01 17:12:57
#> 53 115.8103 -41.79399 2001-01-01 17:14:47
#> 54 115.8204 -41.78278 2001-01-01 17:16:36
#> 55 115.8306 -41.77157 2001-01-01 17:18:26
#> 56 115.8408 -41.76035 2001-01-01 17:20:15
#> 57 115.8509 -41.74914 2001-01-01 17:22:05
#> 58 115.8611 -41.73792 2001-01-01 17:23:54
#> 59 115.8712 -41.72671 2001-01-01 17:25:44
#> 60 115.8814 -41.71549 2001-01-01 17:27:33
#> 61 115.8915 -41.70427 2001-01-01 17:29:23
#> 62 115.9016 -41.69305 2001-01-01 17:31:13
#> 63 115.9118 -41.68183 2001-01-01 17:33:02
#> 64 115.9219 -41.67061 2001-01-01 17:34:52
#> 65 115.9320 -41.65938 2001-01-01 17:36:41
#> 66 115.9422 -41.64816 2001-01-01 17:38:31
#> 67 115.9523 -41.63694 2001-01-01 17:40:20
#> 68 115.9624 -41.62571 2001-01-01 17:42:10
#> 69 115.9725 -41.61449 2001-01-01 17:43:59
#> 70 115.9826 -41.60326 2001-01-01 17:45:49
#> 71 115.9927 -41.59203 2001-01-01 17:47:39
#> 72 116.0028 -41.58080 2001-01-01 17:49:28
#> 73 116.0129 -41.56957 2001-01-01 17:51:18
#> 74 116.0230 -41.55834 2001-01-01 17:53:07
#> 75 116.0331 -41.54711 2001-01-01 17:54:57
#> 76 116.0432 -41.53588 2001-01-01 17:56:46
#> 77 116.0533 -41.52465 2001-01-01 17:58:36
#> 78 116.0634 -41.51341 2001-01-01 18:00:25
#> 79 116.0734 -41.50218 2001-01-01 18:02:15
#> 80 116.0835 -41.49094 2001-01-01 18:04:05
#> 81 116.0936 -41.47971 2001-01-01 18:05:54
#> 82 116.1037 -41.46847 2001-01-01 18:07:44
#> 83 116.1137 -41.45723 2001-01-01 18:09:33
#> 84 116.1238 -41.44599 2001-01-01 18:11:23
#> 85 116.1338 -41.43475 2001-01-01 18:13:12
#> 86 116.1439 -41.42351 2001-01-01 18:15:02
#> 87 116.1539 -41.41227 2001-01-01 18:16:52
#> 
#> $`2`
#>       int_x     int_y            int_date
#> 1  116.1539 -41.41227 2001-01-01 18:16:52
#> 2  116.1459 -41.40013 2001-01-01 18:24:07
#> 3  116.1379 -41.38800 2001-01-01 18:31:22
#> 4  116.1299 -41.37586 2001-01-01 18:38:37
#> 5  116.1219 -41.36372 2001-01-01 18:45:52
#> 6  116.1138 -41.35158 2001-01-01 18:53:07
#> 7  116.1058 -41.33944 2001-01-01 19:00:22
#> 8  116.0978 -41.32730 2001-01-01 19:07:37
#> 9  116.0898 -41.31516 2001-01-01 19:14:52
#> 10 116.0818 -41.30302 2001-01-01 19:22:07
#> 11 116.0738 -41.29088 2001-01-01 19:29:22
#> 12 116.0658 -41.27874 2001-01-01 19:36:37
#> 13 116.0578 -41.26660 2001-01-01 19:43:52
#> 14 116.0498 -41.25445 2001-01-01 19:51:07
#> 15 116.0418 -41.24231 2001-01-01 19:58:22
#> 16 116.0338 -41.23017 2001-01-01 20:05:37
#> 17 116.0259 -41.21802 2001-01-01 20:12:52
#> 18 116.0179 -41.20588 2001-01-01 20:20:07
#> 19 116.0099 -41.19373 2001-01-01 20:27:22
#> 20 116.0019 -41.18158 2001-01-01 20:34:37
#> 21 115.9940 -41.16944 2001-01-01 20:41:52
#> 22 115.9860 -41.15729 2001-01-01 20:49:07
#> 23 115.9780 -41.14514 2001-01-01 20:56:22
#> 24 115.9701 -41.13299 2001-01-01 21:03:38
#> 
#> $`3`
#>       int_x     int_y            int_date
#> 1  115.9701 -41.13299 2001-01-01 21:03:38
#> 2  115.9768 -41.14558 2001-01-01 21:04:27
#> 3  115.9835 -41.15817 2001-01-01 21:05:17
#> 4  115.9902 -41.17076 2001-01-01 21:06:06
#> 5  115.9969 -41.18334 2001-01-01 21:06:56
#> 6  116.0036 -41.19593 2001-01-01 21:07:45
#> 7  116.0103 -41.20852 2001-01-01 21:08:35
#> 8  116.0170 -41.22110 2001-01-01 21:09:24
#> 9  116.0238 -41.23369 2001-01-01 21:10:14
#> 10 116.0305 -41.24628 2001-01-01 21:11:03
#> 11 116.0372 -41.25886 2001-01-01 21:11:53
#> 12 116.0440 -41.27144 2001-01-01 21:12:42
#> 13 116.0507 -41.28403 2001-01-01 21:13:32
#> 14 116.0574 -41.29661 2001-01-01 21:14:21
#> 15 116.0642 -41.30920 2001-01-01 21:15:11
#> 16 116.0709 -41.32178 2001-01-01 21:16:01
#> 17 116.0776 -41.33436 2001-01-01 21:16:50
#> 18 116.0844 -41.34694 2001-01-01 21:17:40
#> 19 116.0911 -41.35953 2001-01-01 21:18:29
#> 20 116.0979 -41.37211 2001-01-01 21:19:19
#> 21 116.1046 -41.38469 2001-01-01 21:20:08
#> 22 116.1114 -41.39727 2001-01-01 21:20:58
#> 23 116.1181 -41.40985 2001-01-01 21:21:47
#> 24 116.1249 -41.42243 2001-01-01 21:22:37
#> 25 116.1317 -41.43501 2001-01-01 21:23:26
#> 26 116.1384 -41.44759 2001-01-01 21:24:16
#> 27 116.1452 -41.46017 2001-01-01 21:25:05
#> 28 116.1520 -41.47274 2001-01-01 21:25:55
#> 29 116.1587 -41.48532 2001-01-01 21:26:45
#> 30 116.1655 -41.49790 2001-01-01 21:27:34
#> 31 116.1723 -41.51048 2001-01-01 21:28:24
#> 32 116.1791 -41.52305 2001-01-01 21:29:13
#> 33 116.1859 -41.53563 2001-01-01 21:30:03
#> 34 116.1927 -41.54821 2001-01-01 21:30:52
#> 35 116.1994 -41.56078 2001-01-01 21:31:42
#> 36 116.2062 -41.57336 2001-01-01 21:32:31
#> 37 116.2130 -41.58593 2001-01-01 21:33:21
#> 38 116.2198 -41.59850 2001-01-01 21:34:10
#> 39 116.2266 -41.61108 2001-01-01 21:35:00
#> 40 116.2334 -41.62365 2001-01-01 21:35:49
#> 41 116.2402 -41.63622 2001-01-01 21:36:39
#> 42 116.2470 -41.64880 2001-01-01 21:37:29
#> 43 116.2539 -41.66137 2001-01-01 21:38:18
#> 44 116.2607 -41.67394 2001-01-01 21:39:08
#> 45 116.2675 -41.68651 2001-01-01 21:39:57
#> 46 116.2743 -41.69908 2001-01-01 21:40:47
#> 47 116.2811 -41.71165 2001-01-01 21:41:36
#> 48 116.2879 -41.72423 2001-01-01 21:42:26
#> 49 116.2948 -41.73679 2001-01-01 21:43:15
#> 50 116.3016 -41.74936 2001-01-01 21:44:05
#> 51 116.3084 -41.76193 2001-01-01 21:44:54
#> 52 116.3153 -41.77450 2001-01-01 21:45:44
#> 53 116.3221 -41.78707 2001-01-01 21:46:33
#> 54 116.3289 -41.79964 2001-01-01 21:47:23
#> 55 116.3358 -41.81221 2001-01-01 21:48:13
#> 56 116.3426 -41.82477 2001-01-01 21:49:02
#> 57 116.3495 -41.83734 2001-01-01 21:49:52
#> 58 116.3563 -41.84991 2001-01-01 21:50:41
#> 59 116.3632 -41.86247 2001-01-01 21:51:31
#> 60 116.3700 -41.87504 2001-01-01 21:52:20
#> 61 116.3769 -41.88760 2001-01-01 21:53:10
#> 62 116.3838 -41.90017 2001-01-01 21:53:59
#> 63 116.3906 -41.91273 2001-01-01 21:54:49
#> 64 116.3975 -41.92529 2001-01-01 21:55:38
#> 65 116.4044 -41.93786 2001-01-01 21:56:28
#> 66 116.4112 -41.95042 2001-01-01 21:57:17
#> 67 116.4181 -41.96298 2001-01-01 21:58:07
#> 68 116.4250 -41.97554 2001-01-01 21:58:57
#> 69 116.4319 -41.98811 2001-01-01 21:59:46
#> 70 116.4388 -42.00067 2001-01-01 22:00:36
#> 71 116.4456 -42.01323 2001-01-01 22:01:25
#> 72 116.4525 -42.02579 2001-01-01 22:02:15
#> 73 116.4594 -42.03835 2001-01-01 22:03:04
#> 74 116.4663 -42.05091 2001-01-01 22:03:54
#> 75 116.4732 -42.06347 2001-01-01 22:04:43
#> 76 116.4801 -42.07603 2001-01-01 22:05:33
#> 77 116.4870 -42.08858 2001-01-01 22:06:22
#> 78 116.4939 -42.10114 2001-01-01 22:07:12
#> 79 116.5008 -42.11370 2001-01-01 22:08:01
#> 80 116.5077 -42.12626 2001-01-01 22:08:51
#> 81 116.5146 -42.13881 2001-01-01 22:09:41
#> 
#> $`4`
#>       int_x     int_y            int_date
#> 1  116.5146 -42.13881 2001-01-01 22:09:41
#> 2  116.5117 -42.12523 2001-01-01 22:13:53
#> 3  116.5087 -42.11165 2001-01-01 22:18:06
#> 4  116.5058 -42.09807 2001-01-01 22:22:18
#> 5  116.5028 -42.08449 2001-01-01 22:26:31
#> 6  116.4999 -42.07091 2001-01-01 22:30:44
#> 7  116.4969 -42.05733 2001-01-01 22:34:56
#> 8  116.4940 -42.04375 2001-01-01 22:39:09
#> 9  116.4910 -42.03017 2001-01-01 22:43:22
#> 10 116.4881 -42.01659 2001-01-01 22:47:34
#> 11 116.4852 -42.00301 2001-01-01 22:51:47
#> 12 116.4822 -41.98943 2001-01-01 22:56:00
#> 13 116.4793 -41.97585 2001-01-01 23:00:12
#> 14 116.4763 -41.96227 2001-01-01 23:04:25
#> 15 116.4734 -41.94869 2001-01-01 23:08:38
#> 16 116.4705 -41.93511 2001-01-01 23:12:50
#> 17 116.4675 -41.92153 2001-01-01 23:17:03
#> 18 116.4646 -41.90794 2001-01-01 23:21:16
#> 19 116.4617 -41.89436 2001-01-01 23:25:28
#> 20 116.4587 -41.88078 2001-01-01 23:29:41
#> 21 116.4558 -41.86720 2001-01-01 23:33:54
#> 
#> $`5`
#>       int_x     int_y            int_date
#> 1  116.4558 -41.86720 2001-01-01 23:33:54
#> 2  116.4737 -41.86919 2001-01-01 23:35:36
#> 3  116.4916 -41.87117 2001-01-01 23:37:19
#> 4  116.5095 -41.87315 2001-01-01 23:39:02
#> 5  116.5274 -41.87513 2001-01-01 23:40:44
#> 6  116.5453 -41.87710 2001-01-01 23:42:27
#> 7  116.5631 -41.87907 2001-01-01 23:44:10
#> 8  116.5810 -41.88104 2001-01-01 23:45:53
#> 9  116.5989 -41.88301 2001-01-01 23:47:35
#> 10 116.6168 -41.88497 2001-01-01 23:49:18
#> 11 116.6347 -41.88693 2001-01-01 23:51:01
#> 12 116.6526 -41.88889 2001-01-01 23:52:44
#> 13 116.6705 -41.89084 2001-01-01 23:54:26
#> 14 116.6884 -41.89280 2001-01-01 23:56:09
#> 15 116.7063 -41.89474 2001-01-01 23:57:52
#> 16 116.7242 -41.89669 2001-01-01 23:59:35
#> 17 116.7421 -41.89864 2001-01-02 00:01:17
#> 18 116.7601 -41.90058 2001-01-02 00:03:00
#> 19 116.7780 -41.90251 2001-01-02 00:04:43
#> 20 116.7959 -41.90445 2001-01-02 00:06:26
#> 21 116.8138 -41.90638 2001-01-02 00:08:08
#> 22 116.8317 -41.90831 2001-01-02 00:09:51
#> 23 116.8496 -41.91024 2001-01-02 00:11:34
#> 24 116.8675 -41.91216 2001-01-02 00:13:16
#> 25 116.8854 -41.91409 2001-01-02 00:14:59
#> 26 116.9033 -41.91600 2001-01-02 00:16:42
#> 27 116.9213 -41.91792 2001-01-02 00:18:25
#> 28 116.9392 -41.91983 2001-01-02 00:20:07
#> 29 116.9571 -41.92174 2001-01-02 00:21:50
#> 30 116.9750 -41.92365 2001-01-02 00:23:33
#> 31 116.9929 -41.92556 2001-01-02 00:25:16
#> 32 117.0109 -41.92746 2001-01-02 00:26:58
#> 33 117.0288 -41.92936 2001-01-02 00:28:41
#> 34 117.0467 -41.93125 2001-01-02 00:30:24
#> 35 117.0646 -41.93315 2001-01-02 00:32:07
#> 36 117.0825 -41.93504 2001-01-02 00:33:49
#> 37 117.1005 -41.93693 2001-01-02 00:35:32
#> 38 117.1184 -41.93881 2001-01-02 00:37:15
#> 39 117.1363 -41.94070 2001-01-02 00:38:58
#> 40 117.1543 -41.94258 2001-01-02 00:40:40
#> 41 117.1722 -41.94445 2001-01-02 00:42:23
#> 42 117.1901 -41.94633 2001-01-02 00:44:06
#> 43 117.2081 -41.94820 2001-01-02 00:45:49
#> 44 117.2260 -41.95007 2001-01-02 00:47:31
#> 45 117.2439 -41.95193 2001-01-02 00:49:14
#> 46 117.2619 -41.95380 2001-01-02 00:50:57
#> 47 117.2798 -41.95566 2001-01-02 00:52:39
#> 48 117.2977 -41.95751 2001-01-02 00:54:22
#> 49 117.3157 -41.95937 2001-01-02 00:56:05
#> 50 117.3336 -41.96122 2001-01-02 00:57:48
#> 51 117.3516 -41.96307 2001-01-02 00:59:30
#> 52 117.3695 -41.96491 2001-01-02 01:01:13
#> 53 117.3874 -41.96676 2001-01-02 01:02:56
#> 54 117.4054 -41.96860 2001-01-02 01:04:39
#> 55 117.4233 -41.97044 2001-01-02 01:06:21
#> 56 117.4413 -41.97227 2001-01-02 01:08:04
#> 57 117.4592 -41.97410 2001-01-02 01:09:47
#> 58 117.4772 -41.97593 2001-01-02 01:11:30
#> 59 117.4951 -41.97776 2001-01-02 01:13:12
#> 60 117.5131 -41.97958 2001-01-02 01:14:55
#> 61 117.5310 -41.98141 2001-01-02 01:16:38
#> 62 117.5490 -41.98322 2001-01-02 01:18:21
#> 63 117.5669 -41.98504 2001-01-02 01:20:03
#> 64 117.5849 -41.98685 2001-01-02 01:21:46
#> 65 117.6028 -41.98866 2001-01-02 01:23:29
#> 66 117.6208 -41.99047 2001-01-02 01:25:12
#> 
#> $`6`
#>       int_x     int_y            int_date
#> 1  117.6208 -41.99047 2001-01-02 01:25:12
#> 2  117.6057 -41.98306 2001-01-02 01:33:38
#> 3  117.5905 -41.97565 2001-01-02 01:42:04
#> 4  117.5754 -41.96823 2001-01-02 01:50:30
#> 5  117.5603 -41.96081 2001-01-02 01:58:56
#> 6  117.5452 -41.95340 2001-01-02 02:07:22
#> 7  117.5300 -41.94598 2001-01-02 02:15:48
#> 8  117.5149 -41.93855 2001-01-02 02:24:14
#> 9  117.4998 -41.93113 2001-01-02 02:32:40
#> 10 117.4847 -41.92370 2001-01-02 02:41:06
#> 11 117.4696 -41.91627 2001-01-02 02:49:32
#> 12 117.4545 -41.90884 2001-01-02 02:57:58
#> 13 117.4394 -41.90141 2001-01-02 03:06:24
#> 14 117.4243 -41.89398 2001-01-02 03:14:50
#> 15 117.4092 -41.88654 2001-01-02 03:23:16
#> 16 117.3941 -41.87910 2001-01-02 03:31:42
#> 17 117.3791 -41.87166 2001-01-02 03:40:08
#> 18 117.3640 -41.86422 2001-01-02 03:48:34
#> 19 117.3489 -41.85678 2001-01-02 03:57:00
#> 20 117.3338 -41.84933 2001-01-02 04:05:26
#> 21 117.3188 -41.84188 2001-01-02 04:13:52
#> 22 117.3037 -41.83443 2001-01-02 04:22:18
#> 23 117.2886 -41.82698 2001-01-02 04:30:44
#> 24 117.2736 -41.81952 2001-01-02 04:39:10
#> 25 117.2585 -41.81207 2001-01-02 04:47:36
#> 26 117.2435 -41.80461 2001-01-02 04:56:02
#> 27 117.2284 -41.79715 2001-01-02 05:04:28
#> 28 117.2134 -41.78969 2001-01-02 05:12:54
#> 29 117.1984 -41.78222 2001-01-02 05:21:20
#> 30 117.1833 -41.77476 2001-01-02 05:29:46
#> 31 117.1683 -41.76729 2001-01-02 05:38:12
#> 32 117.1533 -41.75982 2001-01-02 05:46:38
#> 33 117.1382 -41.75235 2001-01-02 05:55:04
#> 34 117.1232 -41.74488 2001-01-02 06:03:30
#> 35 117.1082 -41.73740 2001-01-02 06:11:56
#> 36 117.0932 -41.72992 2001-01-02 06:20:22
#> 37 117.0782 -41.72244 2001-01-02 06:28:48
#> 38 117.0632 -41.71496 2001-01-02 06:37:14
#> 39 117.0481 -41.70748 2001-01-02 06:45:40
#> 
#> $`7`
#>       int_x     int_y            int_date
#> 1  117.0481 -41.70748 2001-01-02 06:45:40
#> 2  117.0583 -41.69621 2001-01-02 06:48:23
#> 3  117.0685 -41.68495 2001-01-02 06:51:06
#> 4  117.0787 -41.67368 2001-01-02 06:53:49
#> 5  117.0889 -41.66241 2001-01-02 06:56:32
#> 6  117.0990 -41.65115 2001-01-02 06:59:15
#> 7  117.1092 -41.63988 2001-01-02 07:01:58
#> 8  117.1194 -41.62861 2001-01-02 07:04:41
#> 9  117.1295 -41.61734 2001-01-02 07:07:25
#> 10 117.1397 -41.60606 2001-01-02 07:10:08
#> 11 117.1499 -41.59479 2001-01-02 07:12:51
#> 12 117.1600 -41.58352 2001-01-02 07:15:34
#> 13 117.1702 -41.57224 2001-01-02 07:18:17
#> 14 117.1803 -41.56097 2001-01-02 07:21:00
#> 15 117.1904 -41.54969 2001-01-02 07:23:43
#> 16 117.2006 -41.53842 2001-01-02 07:26:27
#> 17 117.2107 -41.52714 2001-01-02 07:29:10
#> 18 117.2208 -41.51586 2001-01-02 07:31:53
#> 19 117.2310 -41.50458 2001-01-02 07:34:36
#> 20 117.2411 -41.49330 2001-01-02 07:37:19
#> 21 117.2512 -41.48202 2001-01-02 07:40:02
#> 22 117.2613 -41.47074 2001-01-02 07:42:45
#> 23 117.2714 -41.45945 2001-01-02 07:45:29
#> 24 117.2815 -41.44817 2001-01-02 07:48:12
#> 25 117.2917 -41.43688 2001-01-02 07:50:55
#> 26 117.3018 -41.42560 2001-01-02 07:53:38
#> 27 117.3119 -41.41431 2001-01-02 07:56:21
#> 28 117.3219 -41.40302 2001-01-02 07:59:04
#> 29 117.3320 -41.39173 2001-01-02 08:01:47
#> 30 117.3421 -41.38044 2001-01-02 08:04:31
#> 31 117.3522 -41.36915 2001-01-02 08:07:14
#> 32 117.3623 -41.35786 2001-01-02 08:09:57
#> 33 117.3724 -41.34657 2001-01-02 08:12:40
#> 34 117.3824 -41.33528 2001-01-02 08:15:23
#> 35 117.3925 -41.32398 2001-01-02 08:18:06
#> 36 117.4026 -41.31269 2001-01-02 08:20:49
#> 37 117.4126 -41.30139 2001-01-02 08:23:33
#> 38 117.4227 -41.29010 2001-01-02 08:26:16
#> 39 117.4328 -41.27880 2001-01-02 08:28:59
#> 40 117.4428 -41.26750 2001-01-02 08:31:42
#> 41 117.4529 -41.25620 2001-01-02 08:34:25
#> 42 117.4629 -41.24490 2001-01-02 08:37:08
#> 43 117.4729 -41.23360 2001-01-02 08:39:51
#> 44 117.4830 -41.22230 2001-01-02 08:42:34
#> 45 117.4930 -41.21100 2001-01-02 08:45:18
#> 46 117.5031 -41.19970 2001-01-02 08:48:01
#> 47 117.5131 -41.18839 2001-01-02 08:50:44
#> 48 117.5231 -41.17709 2001-01-02 08:53:27
#> 49 117.5331 -41.16578 2001-01-02 08:56:10
#> 50 117.5431 -41.15447 2001-01-02 08:58:53
#> 51 117.5532 -41.14317 2001-01-02 09:01:36
#> 52 117.5632 -41.13186 2001-01-02 09:04:20
#> 53 117.5732 -41.12055 2001-01-02 09:07:03
#> 54 117.5832 -41.10924 2001-01-02 09:09:46
#> 55 117.5932 -41.09793 2001-01-02 09:12:29
#> 56 117.6032 -41.08661 2001-01-02 09:15:12
#> 57 117.6132 -41.07530 2001-01-02 09:17:55
#> 58 117.6232 -41.06399 2001-01-02 09:20:38
#> 59 117.6332 -41.05267 2001-01-02 09:23:22
#> 60 117.6431 -41.04136 2001-01-02 09:26:05
#> 61 117.6531 -41.03004 2001-01-02 09:28:48
#> 62 117.6631 -41.01872 2001-01-02 09:31:31
#> 63 117.6731 -41.00740 2001-01-02 09:34:14
#> 64 117.6830 -40.99609 2001-01-02 09:36:57
#> 65 117.6930 -40.98477 2001-01-02 09:39:40
#> 66 117.7030 -40.97345 2001-01-02 09:42:24
#> 67 117.7129 -40.96212 2001-01-02 09:45:07
#> 68 117.7229 -40.95080 2001-01-02 09:47:50
#> 69 117.7328 -40.93948 2001-01-02 09:50:33
#> 70 117.7428 -40.92815 2001-01-02 09:53:16
#> 71 117.7527 -40.91683 2001-01-02 09:55:59
#> 72 117.7627 -40.90550 2001-01-02 09:58:42
#> 73 117.7726 -40.89418 2001-01-02 10:01:26
#> 
#> $`8`
#>       int_x     int_y            int_date
#> 1  117.7726 -40.89418 2001-01-02 10:01:26
#> 2  117.7766 -40.88100 2001-01-02 10:04:03
#> 3  117.7806 -40.86782 2001-01-02 10:06:41
#> 4  117.7845 -40.85464 2001-01-02 10:09:18
#> 5  117.7885 -40.84147 2001-01-02 10:11:56
#> 6  117.7925 -40.82829 2001-01-02 10:14:34
#> 7  117.7964 -40.81511 2001-01-02 10:17:11
#> 8  117.8004 -40.80193 2001-01-02 10:19:49
#> 9  117.8043 -40.78875 2001-01-02 10:22:26
#> 10 117.8083 -40.77558 2001-01-02 10:25:04
#> 11 117.8123 -40.76240 2001-01-02 10:27:42
#> 12 117.8162 -40.74922 2001-01-02 10:30:19
#> 13 117.8202 -40.73604 2001-01-02 10:32:57
#> 14 117.8241 -40.72286 2001-01-02 10:35:35
#> 15 117.8281 -40.70968 2001-01-02 10:38:12
#> 16 117.8320 -40.69650 2001-01-02 10:40:50
#> 17 117.8360 -40.68332 2001-01-02 10:43:27
#> 18 117.8399 -40.67014 2001-01-02 10:46:05
#> 19 117.8439 -40.65696 2001-01-02 10:48:43
#> 20 117.8478 -40.64378 2001-01-02 10:51:20
#> 21 117.8517 -40.63060 2001-01-02 10:53:58
#> 22 117.8557 -40.61742 2001-01-02 10:56:36
#> 23 117.8596 -40.60424 2001-01-02 10:59:13
#> 24 117.8636 -40.59106 2001-01-02 11:01:51
#> 25 117.8675 -40.57788 2001-01-02 11:04:28
#> 26 117.8714 -40.56470 2001-01-02 11:07:06
#> 27 117.8754 -40.55152 2001-01-02 11:09:44
#> 28 117.8793 -40.53834 2001-01-02 11:12:21
#> 29 117.8832 -40.52516 2001-01-02 11:14:59
#> 30 117.8871 -40.51198 2001-01-02 11:17:37
#> 31 117.8911 -40.49880 2001-01-02 11:20:14
#> 32 117.8950 -40.48562 2001-01-02 11:22:52
#> 33 117.8989 -40.47243 2001-01-02 11:25:29
#> 34 117.9028 -40.45925 2001-01-02 11:28:07
#> 35 117.9068 -40.44607 2001-01-02 11:30:45
#> 36 117.9107 -40.43289 2001-01-02 11:33:22
#> 37 117.9146 -40.41971 2001-01-02 11:36:00
#> 38 117.9185 -40.40652 2001-01-02 11:38:37
#> 39 117.9224 -40.39334 2001-01-02 11:41:15
#> 40 117.9263 -40.38016 2001-01-02 11:43:53
#> 41 117.9302 -40.36698 2001-01-02 11:46:30
#> 42 117.9342 -40.35379 2001-01-02 11:49:08
#> 43 117.9381 -40.34061 2001-01-02 11:51:46
#> 44 117.9420 -40.32743 2001-01-02 11:54:23
#> 45 117.9459 -40.31424 2001-01-02 11:57:01
#> 46 117.9498 -40.30106 2001-01-02 11:59:38
#> 47 117.9537 -40.28788 2001-01-02 12:02:16
#> 48 117.9576 -40.27469 2001-01-02 12:04:54
#> 49 117.9615 -40.26151 2001-01-02 12:07:31
#> 50 117.9654 -40.24833 2001-01-02 12:10:09
#> 51 117.9693 -40.23514 2001-01-02 12:12:47
#> 52 117.9732 -40.22196 2001-01-02 12:15:24
#> 53 117.9771 -40.20878 2001-01-02 12:18:02
#> 54 117.9809 -40.19559 2001-01-02 12:20:39
#> 55 117.9848 -40.18241 2001-01-02 12:23:17
#> 56 117.9887 -40.16922 2001-01-02 12:25:55
#> 57 117.9926 -40.15604 2001-01-02 12:28:32
#> 58 117.9965 -40.14285 2001-01-02 12:31:10
#> 59 118.0004 -40.12967 2001-01-02 12:33:48
#> 60 118.0043 -40.11648 2001-01-02 12:36:25
#> 61 118.0081 -40.10330 2001-01-02 12:39:03
#> 62 118.0120 -40.09011 2001-01-02 12:41:40
#> 63 118.0159 -40.07693 2001-01-02 12:44:18
#> 64 118.0198 -40.06374 2001-01-02 12:46:56
#> 65 118.0236 -40.05056 2001-01-02 12:49:33
#> 66 118.0275 -40.03737 2001-01-02 12:52:11
#> 67 118.0314 -40.02418 2001-01-02 12:54:48
#> 68 118.0353 -40.01100 2001-01-02 12:57:26
#> 69 118.0391 -39.99781 2001-01-02 13:00:04
#> 70 118.0430 -39.98462 2001-01-02 13:02:41
#> 71 118.0469 -39.97144 2001-01-02 13:05:19
#> 72 118.0507 -39.95825 2001-01-02 13:07:57
#> 73 118.0546 -39.94506 2001-01-02 13:10:34
#> 74 118.0584 -39.93188 2001-01-02 13:13:12
#> 75 118.0623 -39.91869 2001-01-02 13:15:49
#> 76 118.0662 -39.90550 2001-01-02 13:18:27
#> 77 118.0700 -39.89232 2001-01-02 13:21:05
#> 78 118.0739 -39.87913 2001-01-02 13:23:42
#> 79 118.0777 -39.86594 2001-01-02 13:26:20
#> 80 118.0816 -39.85275 2001-01-02 13:28:58
#> 81 118.0854 -39.83957 2001-01-02 13:31:35
#> 82 118.0893 -39.82638 2001-01-02 13:34:13
#> 83 118.0931 -39.81319 2001-01-02 13:36:50
#> 84 118.0970 -39.80000 2001-01-02 13:39:28
#> 85 118.1008 -39.78681 2001-01-02 13:42:06
#> 86 118.1047 -39.77363 2001-01-02 13:44:43
#> 87 118.1085 -39.76044 2001-01-02 13:47:21
#> 88 118.1123 -39.74725 2001-01-02 13:49:59
#> 
#> $`9`
#>       int_x     int_y            int_date
#> 1  118.1123 -39.74725 2001-01-02 13:49:59
#> 2  118.1157 -39.76059 2001-01-02 13:52:41
#> 3  118.1191 -39.77394 2001-01-02 13:55:24
#> 4  118.1225 -39.78728 2001-01-02 13:58:07
#> 5  118.1259 -39.80062 2001-01-02 14:00:50
#> 6  118.1293 -39.81397 2001-01-02 14:03:33
#> 7  118.1327 -39.82731 2001-01-02 14:06:16
#> 8  118.1361 -39.84065 2001-01-02 14:08:59
#> 9  118.1395 -39.85400 2001-01-02 14:11:42
#> 10 118.1429 -39.86734 2001-01-02 14:14:25
#> 11 118.1463 -39.88068 2001-01-02 14:17:08
#> 12 118.1497 -39.89403 2001-01-02 14:19:51
#> 13 118.1532 -39.90737 2001-01-02 14:22:34
#> 14 118.1566 -39.92071 2001-01-02 14:25:17
#> 15 118.1600 -39.93405 2001-01-02 14:28:00
#> 16 118.1634 -39.94740 2001-01-02 14:30:42
#> 17 118.1668 -39.96074 2001-01-02 14:33:25
#> 18 118.1702 -39.97408 2001-01-02 14:36:08
#> 19 118.1736 -39.98742 2001-01-02 14:38:51
#> 20 118.1770 -40.00076 2001-01-02 14:41:34
#> 21 118.1805 -40.01411 2001-01-02 14:44:17
#> 22 118.1839 -40.02745 2001-01-02 14:47:00
#> 23 118.1873 -40.04079 2001-01-02 14:49:43
#> 24 118.1907 -40.05413 2001-01-02 14:52:26
#> 25 118.1942 -40.06747 2001-01-02 14:55:09
#> 26 118.1976 -40.08082 2001-01-02 14:57:52
#> 27 118.2010 -40.09416 2001-01-02 15:00:35
#> 28 118.2044 -40.10750 2001-01-02 15:03:18
#> 29 118.2079 -40.12084 2001-01-02 15:06:01
#> 30 118.2113 -40.13418 2001-01-02 15:08:43
#> 31 118.2147 -40.14752 2001-01-02 15:11:26
#> 32 118.2182 -40.16086 2001-01-02 15:14:09
#> 33 118.2216 -40.17420 2001-01-02 15:16:52
#> 34 118.2250 -40.18754 2001-01-02 15:19:35
#> 35 118.2285 -40.20088 2001-01-02 15:22:18
#> 36 118.2319 -40.21422 2001-01-02 15:25:01
#> 37 118.2353 -40.22756 2001-01-02 15:27:44
#> 38 118.2388 -40.24090 2001-01-02 15:30:27
#> 39 118.2422 -40.25425 2001-01-02 15:33:10
#> 40 118.2457 -40.26759 2001-01-02 15:35:53
#> 41 118.2491 -40.28093 2001-01-02 15:38:36
#> 42 118.2526 -40.29426 2001-01-02 15:41:19
#> 43 118.2560 -40.30760 2001-01-02 15:44:02
#> 44 118.2595 -40.32094 2001-01-02 15:46:44
#> 45 118.2629 -40.33428 2001-01-02 15:49:27
#> 46 118.2664 -40.34762 2001-01-02 15:52:10
#> 47 118.2698 -40.36096 2001-01-02 15:54:53
#> 48 118.2733 -40.37430 2001-01-02 15:57:36
#> 49 118.2767 -40.38764 2001-01-02 16:00:19
#> 50 118.2802 -40.40098 2001-01-02 16:03:02
#> 51 118.2836 -40.41432 2001-01-02 16:05:45
#> 52 118.2871 -40.42766 2001-01-02 16:08:28
#> 53 118.2906 -40.44100 2001-01-02 16:11:11
#> 54 118.2940 -40.45434 2001-01-02 16:13:54
#> 55 118.2975 -40.46767 2001-01-02 16:16:37
#> 56 118.3010 -40.48101 2001-01-02 16:19:20
#> 57 118.3044 -40.49435 2001-01-02 16:22:03
#> 58 118.3079 -40.50769 2001-01-02 16:24:46
#> 
#> [[10]]
#> data frame with 0 columns and 0 rows
#> 

inter_time <- track_intermediate(trips0$x[1:10], trips0$y[1:10],
                            date = trips0$date, duration = 1800)

#> Error in .vectorize_interm(cbind(x[-n], y[-n]), cbind(x[-1], y[-1]), n = npoints): length(n) == 1L || length(n) == nrow(p1) is not TRUE
# NOT RUN {
## run with full workflow to expand into a new data frame with
## `int_x`, `int_y`, and (optional) `int_date`
 if (requireNamespace("tidyr") && requireNamespace("dplyr")) {
 tr1 <- trips0[seq(1, nrow(trips0), by = 30), ]
   dd <- tr1 %>% group_by(id) %>%
     mutate(inter = track_intermediate(x, y, date = date, distance = 150000)) %>%
     tidyr::unnest()
   plot(dd$int_date, dd$int_y, pch = ".", cex = 2, main = "equidistant in space")
   abline(v = tr1$date)

   dd1 <- tr1 %>% group_by(id) %>%
     mutate(inter = track_intermediate(x, y, date = date, duration = 3600 * 12)) %>%
     tidyr::unnest()
   plot(dd1$int_date, dd1$int_y, pch = ".", cex = 2, main = "equispaced in time")
   abline(v = tr1$date)
 }
# }