Track intermediate points

Calculate great circle intermediate points on longitude, latitude input vectors. A spherical model is used, from the geosphere package.

track_intermediate(x, y, date = NULL, distance = NULL, duration = NULL)

Arguments

x

longitude

y

latitude

date

optional input date-time in POSIXct

distance

optional minimum distance (metres) between interpolated points

duration

optional minimum duration (seconds) between interpolated point, if set then distance must be NULL and date must be input

Value

a list of data frames of intermediate points (for use with unnest() from tidyr)

Details

This function returns a list of data frames, with a data frame of interpolated locations for every interval between input locations. There is a final empty data frame to ensure the list is the same length as the inputs. See embedded usage of the tidyr function 'unnest()' for ease of use.

To use on multiple track ids, use a grouped data frame with tidyverse code like inter <- data %>% group_by(id) %>% mutate(inter = track_intermediate(lon, lat, date = , distance = ).

Then, un-nest this result for further use (the 'inter' above retains the information about the parent locations for custom usage if needed), so the final location of each group has invalid intermediates: dd <- inter %>% slice(-1) %>% unnest()

Examples

track_intermediate(trips0$x[1:10], trips0$y[1:10], distance = 15000)
#> [[1]]
#>       int_x     int_y
#> 1  115.2767 -42.37580
#> 2  115.3656 -42.27975
#> 3  115.4542 -42.18363
#> 4  115.5426 -42.08744
#> 5  115.6308 -41.99119
#> 6  115.7186 -41.89487
#> 7  115.8062 -41.79848
#> 8  115.8935 -41.70203
#> 9  115.9806 -41.60550
#> 10 116.0674 -41.50892
#> 11 116.1539 -41.41227
#> 
#> [[2]]
#>      int_x     int_y
#> 1 116.1539 -41.41227
#> 2 116.1078 -41.34248
#> 3 116.0618 -41.27267
#> 4 116.0159 -41.20284
#> 5 115.9701 -41.13299
#> 
#> [[3]]
#>       int_x     int_y
#> 1  115.9701 -41.13299
#> 2  116.0238 -41.23369
#> 3  116.0776 -41.33436
#> 4  116.1317 -41.43501
#> 5  116.1859 -41.53563
#> 6  116.2402 -41.63622
#> 7  116.2948 -41.73679
#> 8  116.3495 -41.83734
#> 9  116.4044 -41.93786
#> 10 116.4594 -42.03835
#> 11 116.5146 -42.13881
#> 
#> [[4]]
#>      int_x     int_y
#> 1 116.5146 -42.13881
#> 2 116.4999 -42.07091
#> 3 116.4852 -42.00301
#> 4 116.4705 -41.93511
#> 5 116.4558 -41.86720
#> 
#> [[5]]
#>      int_x     int_y
#> 1 116.4558 -41.86720
#> 2 116.6012 -41.88325
#> 3 116.7466 -41.89912
#> 4 116.8921 -41.91481
#> 5 117.0377 -41.93031
#> 6 117.1834 -41.94562
#> 7 117.3291 -41.96076
#> 8 117.4749 -41.97571
#> 9 117.6208 -41.99047
#> 
#> [[6]]
#>      int_x     int_y
#> 1 117.6208 -41.99047
#> 2 117.5059 -41.93410
#> 3 117.3911 -41.87761
#> 4 117.2766 -41.82102
#> 5 117.1623 -41.76430
#> 6 117.0481 -41.70748
#> 
#> [[7]]
#>       int_x     int_y
#> 1  117.0481 -41.70748
#> 2  117.1295 -41.61734
#> 3  117.2107 -41.52714
#> 4  117.2917 -41.43688
#> 5  117.3724 -41.34657
#> 6  117.4529 -41.25620
#> 7  117.5331 -41.16578
#> 8  117.6132 -41.07530
#> 9  117.6930 -40.98477
#> 10 117.7726 -40.89418
#> 
#> [[8]]
#>       int_x     int_y
#> 1  117.7726 -40.89418
#> 2  117.8071 -40.77953
#> 3  117.8415 -40.66487
#> 4  117.8758 -40.55020
#> 5  117.9099 -40.43552
#> 6  117.9439 -40.32084
#> 7  117.9778 -40.20614
#> 8  118.0116 -40.09143
#> 9  118.0453 -39.97671
#> 10 118.0789 -39.86199
#> 11 118.1123 -39.74725
#> 
#> [[9]]
#>      int_x     int_y
#> 1 118.1123 -39.74725
#> 2 118.1400 -39.85590
#> 3 118.1678 -39.96455
#> 4 118.1956 -40.07319
#> 5 118.2236 -40.18183
#> 6 118.2516 -40.29045
#> 7 118.2797 -40.39907
#> 8 118.3079 -40.50769
#> 
#> [[10]]
#> data frame with 0 columns and 0 rows
#> 

track_intermediate(trips0$x[1:10], trips0$y[1:10], date = trips0$date,
                                 distance = 1500)
#> [[1]]
#>       int_x     int_y            int_date
#> 1  115.2767 -42.37580 2001-01-01 15:39:50
#> 2  115.2868 -42.36489 2001-01-01 15:41:37
#> 3  115.2969 -42.35397 2001-01-01 15:43:24
#> 4  115.3070 -42.34306 2001-01-01 15:45:11
#> 5  115.3171 -42.33215 2001-01-01 15:46:58
#> 6  115.3272 -42.32123 2001-01-01 15:48:45
#> 7  115.3373 -42.31032 2001-01-01 15:50:32
#> 8  115.3474 -42.29940 2001-01-01 15:52:19
#> 9  115.3575 -42.28848 2001-01-01 15:54:06
#> 10 115.3676 -42.27756 2001-01-01 15:55:53
#> 11 115.3777 -42.26664 2001-01-01 15:57:40
#> 12 115.3878 -42.25572 2001-01-01 15:59:27
#> 13 115.3979 -42.24480 2001-01-01 16:01:14
#> 14 115.4079 -42.23388 2001-01-01 16:03:01
#> 15 115.4180 -42.22296 2001-01-01 16:04:48
#> 16 115.4281 -42.21203 2001-01-01 16:06:36
#> 17 115.4381 -42.20111 2001-01-01 16:08:23
#> 18 115.4482 -42.19018 2001-01-01 16:10:10
#> 19 115.4583 -42.17926 2001-01-01 16:11:57
#> 20 115.4683 -42.16833 2001-01-01 16:13:44
#> 21 115.4784 -42.15740 2001-01-01 16:15:31
#> 22 115.4884 -42.14647 2001-01-01 16:17:18
#> 23 115.4985 -42.13554 2001-01-01 16:19:05
#> 24 115.5085 -42.12461 2001-01-01 16:20:52
#> 25 115.5186 -42.11368 2001-01-01 16:22:39
#> 26 115.5286 -42.10275 2001-01-01 16:24:26
#> 27 115.5386 -42.09182 2001-01-01 16:26:13
#> 28 115.5486 -42.08088 2001-01-01 16:28:00
#> 29 115.5587 -42.06995 2001-01-01 16:29:47
#> 30 115.5687 -42.05901 2001-01-01 16:31:34
#> 31 115.5787 -42.04807 2001-01-01 16:33:22
#> 32 115.5887 -42.03714 2001-01-01 16:35:09
#> 33 115.5987 -42.02620 2001-01-01 16:36:56
#> 34 115.6087 -42.01526 2001-01-01 16:38:43
#> 35 115.6188 -42.00432 2001-01-01 16:40:30
#> 36 115.6288 -41.99338 2001-01-01 16:42:17
#> 37 115.6388 -41.98243 2001-01-01 16:44:04
#> 38 115.6487 -41.97149 2001-01-01 16:45:51
#> 39 115.6587 -41.96055 2001-01-01 16:47:38
#> 40 115.6687 -41.94960 2001-01-01 16:49:25
#> 41 115.6787 -41.93866 2001-01-01 16:51:12
#> 42 115.6887 -41.92771 2001-01-01 16:52:59
#> 43 115.6987 -41.91676 2001-01-01 16:54:46
#> 44 115.7086 -41.90582 2001-01-01 16:56:33
#> 45 115.7186 -41.89487 2001-01-01 16:58:21
#> 46 115.7286 -41.88392 2001-01-01 17:00:08
#> 47 115.7385 -41.87297 2001-01-01 17:01:55
#> 48 115.7485 -41.86202 2001-01-01 17:03:42
#> 49 115.7585 -41.85106 2001-01-01 17:05:29
#> 50 115.7684 -41.84011 2001-01-01 17:07:16
#> 51 115.7784 -41.82916 2001-01-01 17:09:03
#> 52 115.7883 -41.81820 2001-01-01 17:10:50
#> 53 115.7982 -41.80724 2001-01-01 17:12:37
#> 54 115.8082 -41.79629 2001-01-01 17:14:24
#> 55 115.8181 -41.78533 2001-01-01 17:16:11
#> 56 115.8281 -41.77437 2001-01-01 17:17:58
#> 57 115.8380 -41.76341 2001-01-01 17:19:45
#> 58 115.8479 -41.75245 2001-01-01 17:21:32
#> 59 115.8578 -41.74149 2001-01-01 17:23:19
#> 60 115.8678 -41.73053 2001-01-01 17:25:07
#> 61 115.8777 -41.71957 2001-01-01 17:26:54
#> 62 115.8876 -41.70860 2001-01-01 17:28:41
#> 63 115.8975 -41.69764 2001-01-01 17:30:28
#> 64 115.9074 -41.68667 2001-01-01 17:32:15
#> 65 115.9173 -41.67571 2001-01-01 17:34:02
#> 66 115.9272 -41.66474 2001-01-01 17:35:49
#> 67 115.9371 -41.65377 2001-01-01 17:37:36
#> 68 115.9470 -41.64280 2001-01-01 17:39:23
#> 69 115.9569 -41.63184 2001-01-01 17:41:10
#> 70 115.9668 -41.62086 2001-01-01 17:42:57
#> 71 115.9766 -41.60989 2001-01-01 17:44:44
#> 72 115.9865 -41.59892 2001-01-01 17:46:31
#> 73 115.9964 -41.58795 2001-01-01 17:48:18
#> 74 116.0063 -41.57698 2001-01-01 17:50:05
#> 75 116.0161 -41.56600 2001-01-01 17:51:53
#> 76 116.0260 -41.55503 2001-01-01 17:53:40
#> 77 116.0359 -41.54405 2001-01-01 17:55:27
#> 78 116.0457 -41.53307 2001-01-01 17:57:14
#> 79 116.0556 -41.52209 2001-01-01 17:59:01
#> 80 116.0654 -41.51112 2001-01-01 18:00:48
#> 81 116.0753 -41.50014 2001-01-01 18:02:35
#> 82 116.0851 -41.48916 2001-01-01 18:04:22
#> 83 116.0950 -41.47817 2001-01-01 18:06:09
#> 84 116.1048 -41.46719 2001-01-01 18:07:56
#> 85 116.1146 -41.45621 2001-01-01 18:09:43
#> 86 116.1245 -41.44522 2001-01-01 18:11:30
#> 87 116.1343 -41.43424 2001-01-01 18:13:17
#> 88 116.1441 -41.42325 2001-01-01 18:15:04
#> 89 116.1539 -41.41227 2001-01-01 18:16:52
#> 
#> [[2]]
#>       int_x     int_y            int_date
#> 1  116.1539 -41.41227 2001-01-01 18:16:52
#> 2  116.1466 -41.40110 2001-01-01 18:23:32
#> 3  116.1392 -41.38994 2001-01-01 18:30:12
#> 4  116.1318 -41.37877 2001-01-01 18:36:52
#> 5  116.1244 -41.36760 2001-01-01 18:43:32
#> 6  116.1170 -41.35644 2001-01-01 18:50:13
#> 7  116.1097 -41.34527 2001-01-01 18:56:53
#> 8  116.1023 -41.33410 2001-01-01 19:03:33
#> 9  116.0949 -41.32293 2001-01-01 19:10:13
#> 10 116.0876 -41.31176 2001-01-01 19:16:54
#> 11 116.0802 -41.30059 2001-01-01 19:23:34
#> 12 116.0728 -41.28942 2001-01-01 19:30:14
#> 13 116.0655 -41.27825 2001-01-01 19:36:54
#> 14 116.0581 -41.26708 2001-01-01 19:43:35
#> 15 116.0508 -41.25591 2001-01-01 19:50:15
#> 16 116.0434 -41.24474 2001-01-01 19:56:55
#> 17 116.0361 -41.23357 2001-01-01 20:03:35
#> 18 116.0287 -41.22239 2001-01-01 20:10:16
#> 19 116.0214 -41.21122 2001-01-01 20:16:56
#> 20 116.0141 -41.20005 2001-01-01 20:23:36
#> 21 116.0067 -41.18887 2001-01-01 20:30:16
#> 22 115.9994 -41.17770 2001-01-01 20:36:57
#> 23 115.9921 -41.16652 2001-01-01 20:43:37
#> 24 115.9847 -41.15535 2001-01-01 20:50:17
#> 25 115.9774 -41.14417 2001-01-01 20:56:57
#> 26 115.9701 -41.13299 2001-01-01 21:03:38
#> 
#> [[3]]
#>       int_x     int_y            int_date
#> 1  115.9701 -41.13299 2001-01-01 21:03:38
#> 2  115.9766 -41.14527 2001-01-01 21:04:26
#> 3  115.9832 -41.15756 2001-01-01 21:05:14
#> 4  115.9897 -41.16984 2001-01-01 21:06:02
#> 5  115.9962 -41.18212 2001-01-01 21:06:51
#> 6  116.0028 -41.19440 2001-01-01 21:07:39
#> 7  116.0093 -41.20668 2001-01-01 21:08:27
#> 8  116.0159 -41.21896 2001-01-01 21:09:16
#> 9  116.0225 -41.23123 2001-01-01 21:10:04
#> 10 116.0290 -41.24351 2001-01-01 21:10:52
#> 11 116.0356 -41.25579 2001-01-01 21:11:41
#> 12 116.0421 -41.26807 2001-01-01 21:12:29
#> 13 116.0487 -41.28035 2001-01-01 21:13:17
#> 14 116.0553 -41.29262 2001-01-01 21:14:06
#> 15 116.0619 -41.30490 2001-01-01 21:14:54
#> 16 116.0684 -41.31718 2001-01-01 21:15:42
#> 17 116.0750 -41.32945 2001-01-01 21:16:31
#> 18 116.0816 -41.34173 2001-01-01 21:17:19
#> 19 116.0882 -41.35400 2001-01-01 21:18:07
#> 20 116.0947 -41.36628 2001-01-01 21:18:56
#> 21 116.1013 -41.37855 2001-01-01 21:19:44
#> 22 116.1079 -41.39083 2001-01-01 21:20:32
#> 23 116.1145 -41.40310 2001-01-01 21:21:21
#> 24 116.1211 -41.41537 2001-01-01 21:22:09
#> 25 116.1277 -41.42764 2001-01-01 21:22:57
#> 26 116.1343 -41.43992 2001-01-01 21:23:46
#> 27 116.1409 -41.45219 2001-01-01 21:24:34
#> 28 116.1475 -41.46446 2001-01-01 21:25:22
#> 29 116.1541 -41.47673 2001-01-01 21:26:11
#> 30 116.1607 -41.48900 2001-01-01 21:26:59
#> 31 116.1673 -41.50127 2001-01-01 21:27:47
#> 32 116.1740 -41.51354 2001-01-01 21:28:36
#> 33 116.1806 -41.52581 2001-01-01 21:29:24
#> 34 116.1872 -41.53808 2001-01-01 21:30:12
#> 35 116.1938 -41.55035 2001-01-01 21:31:01
#> 36 116.2004 -41.56262 2001-01-01 21:31:49
#> 37 116.2071 -41.57489 2001-01-01 21:32:37
#> 38 116.2137 -41.58716 2001-01-01 21:33:26
#> 39 116.2203 -41.59942 2001-01-01 21:34:14
#> 40 116.2270 -41.61169 2001-01-01 21:35:02
#> 41 116.2336 -41.62396 2001-01-01 21:35:51
#> 42 116.2402 -41.63622 2001-01-01 21:36:39
#> 43 116.2469 -41.64849 2001-01-01 21:37:27
#> 44 116.2535 -41.66076 2001-01-01 21:38:16
#> 45 116.2602 -41.67302 2001-01-01 21:39:04
#> 46 116.2668 -41.68529 2001-01-01 21:39:52
#> 47 116.2735 -41.69755 2001-01-01 21:40:41
#> 48 116.2801 -41.70982 2001-01-01 21:41:29
#> 49 116.2868 -41.72208 2001-01-01 21:42:17
#> 50 116.2934 -41.73434 2001-01-01 21:43:06
#> 51 116.3001 -41.74661 2001-01-01 21:43:54
#> 52 116.3068 -41.75887 2001-01-01 21:44:42
#> 53 116.3134 -41.77113 2001-01-01 21:45:31
#> 54 116.3201 -41.78339 2001-01-01 21:46:19
#> 55 116.3268 -41.79565 2001-01-01 21:47:07
#> 56 116.3335 -41.80791 2001-01-01 21:47:56
#> 57 116.3401 -41.82017 2001-01-01 21:48:44
#> 58 116.3468 -41.83243 2001-01-01 21:49:32
#> 59 116.3535 -41.84469 2001-01-01 21:50:21
#> 60 116.3602 -41.85695 2001-01-01 21:51:09
#> 61 116.3669 -41.86921 2001-01-01 21:51:57
#> 62 116.3736 -41.88147 2001-01-01 21:52:46
#> 63 116.3802 -41.89373 2001-01-01 21:53:34
#> 64 116.3869 -41.90599 2001-01-01 21:54:22
#> 65 116.3936 -41.91825 2001-01-01 21:55:11
#> 66 116.4003 -41.93050 2001-01-01 21:55:59
#> 67 116.4070 -41.94276 2001-01-01 21:56:47
#> 68 116.4137 -41.95502 2001-01-01 21:57:36
#> 69 116.4205 -41.96727 2001-01-01 21:58:24
#> 70 116.4272 -41.97953 2001-01-01 21:59:12
#> 71 116.4339 -41.99178 2001-01-01 22:00:01
#> 72 116.4406 -42.00404 2001-01-01 22:00:49
#> 73 116.4473 -42.01629 2001-01-01 22:01:37
#> 74 116.4540 -42.02855 2001-01-01 22:02:26
#> 75 116.4608 -42.04080 2001-01-01 22:03:14
#> 76 116.4675 -42.05305 2001-01-01 22:04:02
#> 77 116.4742 -42.06531 2001-01-01 22:04:51
#> 78 116.4809 -42.07756 2001-01-01 22:05:39
#> 79 116.4877 -42.08981 2001-01-01 22:06:27
#> 80 116.4944 -42.10206 2001-01-01 22:07:16
#> 81 116.5012 -42.11431 2001-01-01 22:08:04
#> 82 116.5079 -42.12656 2001-01-01 22:08:52
#> 83 116.5146 -42.13881 2001-01-01 22:09:41
#> 
#> [[4]]
#>       int_x     int_y            int_date
#> 1  116.5146 -42.13881 2001-01-01 22:09:41
#> 2  116.5120 -42.12647 2001-01-01 22:13:30
#> 3  116.5093 -42.11412 2001-01-01 22:17:20
#> 4  116.5066 -42.10178 2001-01-01 22:21:10
#> 5  116.5039 -42.08943 2001-01-01 22:24:59
#> 6  116.5012 -42.07709 2001-01-01 22:28:49
#> 7  116.4985 -42.06474 2001-01-01 22:32:39
#> 8  116.4959 -42.05239 2001-01-01 22:36:28
#> 9  116.4932 -42.04005 2001-01-01 22:40:18
#> 10 116.4905 -42.02770 2001-01-01 22:44:08
#> 11 116.4878 -42.01536 2001-01-01 22:47:57
#> 12 116.4852 -42.00301 2001-01-01 22:51:47
#> 13 116.4825 -41.99066 2001-01-01 22:55:37
#> 14 116.4798 -41.97832 2001-01-01 22:59:26
#> 15 116.4771 -41.96597 2001-01-01 23:03:16
#> 16 116.4745 -41.95363 2001-01-01 23:07:06
#> 17 116.4718 -41.94128 2001-01-01 23:10:55
#> 18 116.4691 -41.92893 2001-01-01 23:14:45
#> 19 116.4665 -41.91659 2001-01-01 23:18:35
#> 20 116.4638 -41.90424 2001-01-01 23:22:24
#> 21 116.4611 -41.89189 2001-01-01 23:26:14
#> 22 116.4585 -41.87955 2001-01-01 23:30:04
#> 23 116.4558 -41.86720 2001-01-01 23:33:54
#> 
#> [[5]]
#>       int_x     int_y            int_date
#> 1  116.4558 -41.86720 2001-01-01 23:33:54
#> 2  116.4732 -41.86913 2001-01-01 23:35:33
#> 3  116.4905 -41.87105 2001-01-01 23:37:13
#> 4  116.5079 -41.87297 2001-01-01 23:38:53
#> 5  116.5252 -41.87489 2001-01-01 23:40:32
#> 6  116.5426 -41.87681 2001-01-01 23:42:12
#> 7  116.5599 -41.87872 2001-01-01 23:43:52
#> 8  116.5773 -41.88063 2001-01-01 23:45:31
#> 9  116.5947 -41.88254 2001-01-01 23:47:11
#> 10 116.6120 -41.88444 2001-01-01 23:48:51
#> 11 116.6294 -41.88635 2001-01-01 23:50:30
#> 12 116.6468 -41.88825 2001-01-01 23:52:10
#> 13 116.6641 -41.89014 2001-01-01 23:53:50
#> 14 116.6815 -41.89204 2001-01-01 23:55:29
#> 15 116.6989 -41.89393 2001-01-01 23:57:09
#> 16 116.7162 -41.89582 2001-01-01 23:58:49
#> 17 116.7336 -41.89771 2001-01-02 00:00:28
#> 18 116.7510 -41.89959 2001-01-02 00:02:08
#> 19 116.7683 -41.90147 2001-01-02 00:03:48
#> 20 116.7857 -41.90335 2001-01-02 00:05:27
#> 21 116.8031 -41.90523 2001-01-02 00:07:07
#> 22 116.8205 -41.90710 2001-01-02 00:08:47
#> 23 116.8378 -41.90897 2001-01-02 00:10:26
#> 24 116.8552 -41.91084 2001-01-02 00:12:06
#> 25 116.8726 -41.91271 2001-01-02 00:13:46
#> 26 116.8900 -41.91457 2001-01-02 00:15:25
#> 27 116.9074 -41.91643 2001-01-02 00:17:05
#> 28 116.9247 -41.91829 2001-01-02 00:18:45
#> 29 116.9421 -41.92015 2001-01-02 00:20:24
#> 30 116.9595 -41.92200 2001-01-02 00:22:04
#> 31 116.9769 -41.92385 2001-01-02 00:23:44
#> 32 116.9943 -41.92570 2001-01-02 00:25:23
#> 33 117.0117 -41.92754 2001-01-02 00:27:03
#> 34 117.0290 -41.92939 2001-01-02 00:28:43
#> 35 117.0464 -41.93123 2001-01-02 00:30:22
#> 36 117.0638 -41.93306 2001-01-02 00:32:02
#> 37 117.0812 -41.93490 2001-01-02 00:33:42
#> 38 117.0986 -41.93673 2001-01-02 00:35:21
#> 39 117.1160 -41.93856 2001-01-02 00:37:01
#> 40 117.1334 -41.94039 2001-01-02 00:38:41
#> 41 117.1508 -41.94221 2001-01-02 00:40:20
#> 42 117.1682 -41.94403 2001-01-02 00:42:00
#> 43 117.1856 -41.94585 2001-01-02 00:43:40
#> 44 117.2030 -41.94767 2001-01-02 00:45:19
#> 45 117.2204 -41.94948 2001-01-02 00:46:59
#> 46 117.2378 -41.95129 2001-01-02 00:48:39
#> 47 117.2552 -41.95310 2001-01-02 00:50:18
#> 48 117.2726 -41.95491 2001-01-02 00:51:58
#> 49 117.2900 -41.95671 2001-01-02 00:53:38
#> 50 117.3074 -41.95851 2001-01-02 00:55:17
#> 51 117.3248 -41.96031 2001-01-02 00:56:57
#> 52 117.3422 -41.96210 2001-01-02 00:58:37
#> 53 117.3596 -41.96390 2001-01-02 01:00:16
#> 54 117.3770 -41.96569 2001-01-02 01:01:56
#> 55 117.3944 -41.96747 2001-01-02 01:03:36
#> 56 117.4118 -41.96926 2001-01-02 01:05:15
#> 57 117.4292 -41.97104 2001-01-02 01:06:55
#> 58 117.4466 -41.97282 2001-01-02 01:08:35
#> 59 117.4640 -41.97460 2001-01-02 01:10:14
#> 60 117.4815 -41.97637 2001-01-02 01:11:54
#> 61 117.4989 -41.97814 2001-01-02 01:13:34
#> 62 117.5163 -41.97991 2001-01-02 01:15:13
#> 63 117.5337 -41.98168 2001-01-02 01:16:53
#> 64 117.5511 -41.98344 2001-01-02 01:18:33
#> 65 117.5685 -41.98520 2001-01-02 01:20:12
#> 66 117.5860 -41.98696 2001-01-02 01:21:52
#> 67 117.6034 -41.98872 2001-01-02 01:23:32
#> 68 117.6208 -41.99047 2001-01-02 01:25:12
#> 
#> [[6]]
#>       int_x     int_y            int_date
#> 1  117.6208 -41.99047 2001-01-02 01:25:12
#> 2  117.6064 -41.98343 2001-01-02 01:33:12
#> 3  117.5920 -41.97639 2001-01-02 01:41:13
#> 4  117.5777 -41.96934 2001-01-02 01:49:14
#> 5  117.5633 -41.96230 2001-01-02 01:57:14
#> 6  117.5489 -41.95525 2001-01-02 02:05:15
#> 7  117.5346 -41.94820 2001-01-02 02:13:16
#> 8  117.5202 -41.94115 2001-01-02 02:21:16
#> 9  117.5059 -41.93410 2001-01-02 02:29:17
#> 10 117.4915 -41.92704 2001-01-02 02:37:18
#> 11 117.4772 -41.91999 2001-01-02 02:45:19
#> 12 117.4628 -41.91293 2001-01-02 02:53:19
#> 13 117.4485 -41.90587 2001-01-02 03:01:20
#> 14 117.4341 -41.89881 2001-01-02 03:09:21
#> 15 117.4198 -41.89175 2001-01-02 03:17:21
#> 16 117.4055 -41.88468 2001-01-02 03:25:22
#> 17 117.3911 -41.87761 2001-01-02 03:33:23
#> 18 117.3768 -41.87055 2001-01-02 03:41:23
#> 19 117.3625 -41.86347 2001-01-02 03:49:24
#> 20 117.3482 -41.85640 2001-01-02 03:57:25
#> 21 117.3338 -41.84933 2001-01-02 04:05:26
#> 22 117.3195 -41.84225 2001-01-02 04:13:26
#> 23 117.3052 -41.83518 2001-01-02 04:21:27
#> 24 117.2909 -41.82810 2001-01-02 04:29:28
#> 25 117.2766 -41.82102 2001-01-02 04:37:28
#> 26 117.2623 -41.81393 2001-01-02 04:45:29
#> 27 117.2480 -41.80685 2001-01-02 04:53:30
#> 28 117.2337 -41.79976 2001-01-02 05:01:30
#> 29 117.2194 -41.79267 2001-01-02 05:09:31
#> 30 117.2051 -41.78558 2001-01-02 05:17:32
#> 31 117.1908 -41.77849 2001-01-02 05:25:33
#> 32 117.1766 -41.77140 2001-01-02 05:33:33
#> 33 117.1623 -41.76430 2001-01-02 05:41:34
#> 34 117.1480 -41.75721 2001-01-02 05:49:35
#> 35 117.1337 -41.75011 2001-01-02 05:57:35
#> 36 117.1195 -41.74301 2001-01-02 06:05:36
#> 37 117.1052 -41.73590 2001-01-02 06:13:37
#> 38 117.0909 -41.72880 2001-01-02 06:21:37
#> 39 117.0767 -41.72169 2001-01-02 06:29:38
#> 40 117.0624 -41.71459 2001-01-02 06:37:39
#> 41 117.0481 -41.70748 2001-01-02 06:45:40
#> 
#> [[7]]
#>       int_x     int_y            int_date
#> 1  117.0481 -41.70748 2001-01-02 06:45:40
#> 2  117.0581 -41.69652 2001-01-02 06:48:18
#> 3  117.0680 -41.68556 2001-01-02 06:50:57
#> 4  117.0779 -41.67459 2001-01-02 06:53:36
#> 5  117.0878 -41.66363 2001-01-02 06:56:14
#> 6  117.0977 -41.65267 2001-01-02 06:58:53
#> 7  117.1076 -41.64170 2001-01-02 07:01:32
#> 8  117.1175 -41.63074 2001-01-02 07:04:11
#> 9  117.1273 -41.61977 2001-01-02 07:06:49
#> 10 117.1372 -41.60881 2001-01-02 07:09:28
#> 11 117.1471 -41.59784 2001-01-02 07:12:07
#> 12 117.1570 -41.58687 2001-01-02 07:14:46
#> 13 117.1669 -41.57590 2001-01-02 07:17:24
#> 14 117.1767 -41.56493 2001-01-02 07:20:03
#> 15 117.1866 -41.55396 2001-01-02 07:22:42
#> 16 117.1965 -41.54299 2001-01-02 07:25:20
#> 17 117.2063 -41.53202 2001-01-02 07:27:59
#> 18 117.2162 -41.52104 2001-01-02 07:30:38
#> 19 117.2260 -41.51007 2001-01-02 07:33:17
#> 20 117.2359 -41.49909 2001-01-02 07:35:55
#> 21 117.2457 -41.48812 2001-01-02 07:38:34
#> 22 117.2556 -41.47714 2001-01-02 07:41:13
#> 23 117.2654 -41.46616 2001-01-02 07:43:52
#> 24 117.2753 -41.45518 2001-01-02 07:46:30
#> 25 117.2851 -41.44420 2001-01-02 07:49:09
#> 26 117.2949 -41.43322 2001-01-02 07:51:48
#> 27 117.3048 -41.42224 2001-01-02 07:54:26
#> 28 117.3146 -41.41126 2001-01-02 07:57:05
#> 29 117.3244 -41.40028 2001-01-02 07:59:44
#> 30 117.3342 -41.38929 2001-01-02 08:02:23
#> 31 117.3440 -41.37831 2001-01-02 08:05:01
#> 32 117.3538 -41.36732 2001-01-02 08:07:40
#> 33 117.3637 -41.35634 2001-01-02 08:10:19
#> 34 117.3735 -41.34535 2001-01-02 08:12:58
#> 35 117.3833 -41.33436 2001-01-02 08:15:36
#> 36 117.3931 -41.32337 2001-01-02 08:18:15
#> 37 117.4029 -41.31238 2001-01-02 08:20:54
#> 38 117.4126 -41.30139 2001-01-02 08:23:33
#> 39 117.4224 -41.29040 2001-01-02 08:26:11
#> 40 117.4322 -41.27941 2001-01-02 08:28:50
#> 41 117.4420 -41.26842 2001-01-02 08:31:29
#> 42 117.4518 -41.25743 2001-01-02 08:34:07
#> 43 117.4615 -41.24643 2001-01-02 08:36:46
#> 44 117.4713 -41.23544 2001-01-02 08:39:25
#> 45 117.4811 -41.22444 2001-01-02 08:42:04
#> 46 117.4909 -41.21344 2001-01-02 08:44:42
#> 47 117.5006 -41.20244 2001-01-02 08:47:21
#> 48 117.5104 -41.19145 2001-01-02 08:50:00
#> 49 117.5201 -41.18045 2001-01-02 08:52:39
#> 50 117.5299 -41.16945 2001-01-02 08:55:17
#> 51 117.5396 -41.15845 2001-01-02 08:57:56
#> 52 117.5494 -41.14744 2001-01-02 09:00:35
#> 53 117.5591 -41.13644 2001-01-02 09:03:13
#> 54 117.5689 -41.12544 2001-01-02 09:05:52
#> 55 117.5786 -41.11443 2001-01-02 09:08:31
#> 56 117.5883 -41.10343 2001-01-02 09:11:10
#> 57 117.5981 -41.09242 2001-01-02 09:13:48
#> 58 117.6078 -41.08142 2001-01-02 09:16:27
#> 59 117.6175 -41.07041 2001-01-02 09:19:06
#> 60 117.6272 -41.05940 2001-01-02 09:21:45
#> 61 117.6369 -41.04839 2001-01-02 09:24:23
#> 62 117.6466 -41.03738 2001-01-02 09:27:02
#> 63 117.6564 -41.02637 2001-01-02 09:29:41
#> 64 117.6661 -41.01536 2001-01-02 09:32:19
#> 65 117.6758 -41.00435 2001-01-02 09:34:58
#> 66 117.6855 -40.99333 2001-01-02 09:37:37
#> 67 117.6952 -40.98232 2001-01-02 09:40:16
#> 68 117.7049 -40.97130 2001-01-02 09:42:54
#> 69 117.7145 -40.96029 2001-01-02 09:45:33
#> 70 117.7242 -40.94927 2001-01-02 09:48:12
#> 71 117.7339 -40.93825 2001-01-02 09:50:51
#> 72 117.7436 -40.92724 2001-01-02 09:53:29
#> 73 117.7533 -40.91622 2001-01-02 09:56:08
#> 74 117.7629 -40.90520 2001-01-02 09:58:47
#> 75 117.7726 -40.89418 2001-01-02 10:01:26
#> 
#> [[8]]
#>       int_x     int_y            int_date
#> 1  117.7726 -40.89418 2001-01-02 10:01:26
#> 2  117.7765 -40.88129 2001-01-02 10:04:00
#> 3  117.7804 -40.86841 2001-01-02 10:06:34
#> 4  117.7843 -40.85553 2001-01-02 10:09:08
#> 5  117.7881 -40.84265 2001-01-02 10:11:42
#> 6  117.7920 -40.82977 2001-01-02 10:14:16
#> 7  117.7959 -40.81689 2001-01-02 10:16:50
#> 8  117.7998 -40.80400 2001-01-02 10:19:24
#> 9  117.8036 -40.79112 2001-01-02 10:21:58
#> 10 117.8075 -40.77824 2001-01-02 10:24:32
#> 11 117.8114 -40.76536 2001-01-02 10:27:06
#> 12 117.8152 -40.75248 2001-01-02 10:29:40
#> 13 117.8191 -40.73959 2001-01-02 10:32:14
#> 14 117.8230 -40.72671 2001-01-02 10:34:49
#> 15 117.8268 -40.71383 2001-01-02 10:37:23
#> 16 117.8307 -40.70094 2001-01-02 10:39:57
#> 17 117.8346 -40.68806 2001-01-02 10:42:31
#> 18 117.8384 -40.67518 2001-01-02 10:45:05
#> 19 117.8423 -40.66229 2001-01-02 10:47:39
#> 20 117.8461 -40.64941 2001-01-02 10:50:13
#> 21 117.8500 -40.63653 2001-01-02 10:52:47
#> 22 117.8538 -40.62364 2001-01-02 10:55:21
#> 23 117.8577 -40.61076 2001-01-02 10:57:55
#> 24 117.8615 -40.59787 2001-01-02 11:00:29
#> 25 117.8654 -40.58499 2001-01-02 11:03:03
#> 26 117.8692 -40.57211 2001-01-02 11:05:37
#> 27 117.8731 -40.55922 2001-01-02 11:08:12
#> 28 117.8769 -40.54634 2001-01-02 11:10:46
#> 29 117.8807 -40.53345 2001-01-02 11:13:20
#> 30 117.8846 -40.52057 2001-01-02 11:15:54
#> 31 117.8884 -40.50768 2001-01-02 11:18:28
#> 32 117.8923 -40.49480 2001-01-02 11:21:02
#> 33 117.8961 -40.48191 2001-01-02 11:23:36
#> 34 117.8999 -40.46903 2001-01-02 11:26:10
#> 35 117.9038 -40.45614 2001-01-02 11:28:44
#> 36 117.9076 -40.44326 2001-01-02 11:31:18
#> 37 117.9114 -40.43037 2001-01-02 11:33:52
#> 38 117.9153 -40.41748 2001-01-02 11:36:26
#> 39 117.9191 -40.40460 2001-01-02 11:39:00
#> 40 117.9229 -40.39171 2001-01-02 11:41:35
#> 41 117.9267 -40.37883 2001-01-02 11:44:09
#> 42 117.9306 -40.36594 2001-01-02 11:46:43
#> 43 117.9344 -40.35305 2001-01-02 11:49:17
#> 44 117.9382 -40.34017 2001-01-02 11:51:51
#> 45 117.9420 -40.32728 2001-01-02 11:54:25
#> 46 117.9458 -40.31439 2001-01-02 11:56:59
#> 47 117.9496 -40.30151 2001-01-02 11:59:33
#> 48 117.9535 -40.28862 2001-01-02 12:02:07
#> 49 117.9573 -40.27573 2001-01-02 12:04:41
#> 50 117.9611 -40.26284 2001-01-02 12:07:15
#> 51 117.9649 -40.24996 2001-01-02 12:09:49
#> 52 117.9687 -40.23707 2001-01-02 12:12:24
#> 53 117.9725 -40.22418 2001-01-02 12:14:58
#> 54 117.9763 -40.21129 2001-01-02 12:17:32
#> 55 117.9801 -40.19841 2001-01-02 12:20:06
#> 56 117.9839 -40.18552 2001-01-02 12:22:40
#> 57 117.9877 -40.17263 2001-01-02 12:25:14
#> 58 117.9915 -40.15974 2001-01-02 12:27:48
#> 59 117.9953 -40.14685 2001-01-02 12:30:22
#> 60 117.9991 -40.13396 2001-01-02 12:32:56
#> 61 118.0029 -40.12108 2001-01-02 12:35:30
#> 62 118.0067 -40.10819 2001-01-02 12:38:04
#> 63 118.0105 -40.09530 2001-01-02 12:40:38
#> 64 118.0143 -40.08241 2001-01-02 12:43:12
#> 65 118.0181 -40.06952 2001-01-02 12:45:47
#> 66 118.0219 -40.05663 2001-01-02 12:48:21
#> 67 118.0256 -40.04374 2001-01-02 12:50:55
#> 68 118.0294 -40.03085 2001-01-02 12:53:29
#> 69 118.0332 -40.01796 2001-01-02 12:56:03
#> 70 118.0370 -40.00507 2001-01-02 12:58:37
#> 71 118.0408 -39.99218 2001-01-02 13:01:11
#> 72 118.0446 -39.97929 2001-01-02 13:03:45
#> 73 118.0483 -39.96640 2001-01-02 13:06:19
#> 74 118.0521 -39.95351 2001-01-02 13:08:53
#> 75 118.0559 -39.94062 2001-01-02 13:11:27
#> 76 118.0597 -39.92773 2001-01-02 13:14:01
#> 77 118.0634 -39.91484 2001-01-02 13:16:35
#> 78 118.0672 -39.90195 2001-01-02 13:19:10
#> 79 118.0710 -39.88906 2001-01-02 13:21:44
#> 80 118.0747 -39.87617 2001-01-02 13:24:18
#> 81 118.0785 -39.86327 2001-01-02 13:26:52
#> 82 118.0823 -39.85038 2001-01-02 13:29:26
#> 83 118.0860 -39.83749 2001-01-02 13:32:00
#> 84 118.0898 -39.82460 2001-01-02 13:34:34
#> 85 118.0936 -39.81171 2001-01-02 13:37:08
#> 86 118.0973 -39.79882 2001-01-02 13:39:42
#> 87 118.1011 -39.78592 2001-01-02 13:42:16
#> 88 118.1048 -39.77303 2001-01-02 13:44:50
#> 89 118.1086 -39.76014 2001-01-02 13:47:24
#> 90 118.1123 -39.74725 2001-01-02 13:49:59
#> 
#> [[9]]
#>       int_x     int_y            int_date
#> 1  118.1123 -39.74725 2001-01-02 13:49:59
#> 2  118.1156 -39.76014 2001-01-02 13:52:36
#> 3  118.1189 -39.77303 2001-01-02 13:55:13
#> 4  118.1222 -39.78592 2001-01-02 13:57:51
#> 5  118.1255 -39.79881 2001-01-02 14:00:28
#> 6  118.1287 -39.81171 2001-01-02 14:03:06
#> 7  118.1320 -39.82460 2001-01-02 14:05:43
#> 8  118.1353 -39.83749 2001-01-02 14:08:20
#> 9  118.1386 -39.85038 2001-01-02 14:10:58
#> 10 118.1419 -39.86327 2001-01-02 14:13:35
#> 11 118.1452 -39.87616 2001-01-02 14:16:13
#> 12 118.1485 -39.88905 2001-01-02 14:18:50
#> 13 118.1518 -39.90194 2001-01-02 14:21:27
#> 14 118.1551 -39.91483 2001-01-02 14:24:05
#> 15 118.1584 -39.92772 2001-01-02 14:26:42
#> 16 118.1616 -39.94061 2001-01-02 14:29:20
#> 17 118.1649 -39.95350 2001-01-02 14:31:57
#> 18 118.1682 -39.96639 2001-01-02 14:34:34
#> 19 118.1715 -39.97928 2001-01-02 14:37:12
#> 20 118.1748 -39.99217 2001-01-02 14:39:49
#> 21 118.1781 -40.00506 2001-01-02 14:42:27
#> 22 118.1814 -40.01795 2001-01-02 14:45:04
#> 23 118.1848 -40.03084 2001-01-02 14:47:41
#> 24 118.1881 -40.04373 2001-01-02 14:50:19
#> 25 118.1914 -40.05662 2001-01-02 14:52:56
#> 26 118.1947 -40.06951 2001-01-02 14:55:34
#> 27 118.1980 -40.08240 2001-01-02 14:58:11
#> 28 118.2013 -40.09529 2001-01-02 15:00:48
#> 29 118.2046 -40.10818 2001-01-02 15:03:26
#> 30 118.2079 -40.12107 2001-01-02 15:06:03
#> 31 118.2112 -40.13395 2001-01-02 15:08:41
#> 32 118.2145 -40.14684 2001-01-02 15:11:18
#> 33 118.2179 -40.15973 2001-01-02 15:13:56
#> 34 118.2212 -40.17262 2001-01-02 15:16:33
#> 35 118.2245 -40.18551 2001-01-02 15:19:10
#> 36 118.2278 -40.19840 2001-01-02 15:21:48
#> 37 118.2311 -40.21128 2001-01-02 15:24:25
#> 38 118.2345 -40.22417 2001-01-02 15:27:03
#> 39 118.2378 -40.23706 2001-01-02 15:29:40
#> 40 118.2411 -40.24995 2001-01-02 15:32:17
#> 41 118.2444 -40.26284 2001-01-02 15:34:55
#> 42 118.2478 -40.27572 2001-01-02 15:37:32
#> 43 118.2511 -40.28861 2001-01-02 15:40:10
#> 44 118.2544 -40.30150 2001-01-02 15:42:47
#> 45 118.2578 -40.31439 2001-01-02 15:45:24
#> 46 118.2611 -40.32727 2001-01-02 15:48:02
#> 47 118.2644 -40.34016 2001-01-02 15:50:39
#> 48 118.2678 -40.35305 2001-01-02 15:53:17
#> 49 118.2711 -40.36594 2001-01-02 15:55:54
#> 50 118.2744 -40.37882 2001-01-02 15:58:31
#> 51 118.2778 -40.39171 2001-01-02 16:01:09
#> 52 118.2811 -40.40460 2001-01-02 16:03:46
#> 53 118.2845 -40.41748 2001-01-02 16:06:24
#> 54 118.2878 -40.43037 2001-01-02 16:09:01
#> 55 118.2912 -40.44326 2001-01-02 16:11:38
#> 56 118.2945 -40.45614 2001-01-02 16:14:16
#> 57 118.2979 -40.46903 2001-01-02 16:16:53
#> 58 118.3012 -40.48192 2001-01-02 16:19:31
#> 59 118.3046 -40.49480 2001-01-02 16:22:08
#> 60 118.3079 -40.50769 2001-01-02 16:24:46
#> 
#> [[10]]
#> data frame with 0 columns and 0 rows
#> 

inter_time <- track_intermediate(trips0$x[1:10], trips0$y[1:10],
                            date = trips0$date, duration = 1800)
#> Warning: number of rows of result is not a multiple of vector length (arg 1)